Implement LexiGuess API with new services for word generation, caching, and validation; replace LexiGuessController with a modular approach using WordService, CompositeWordGenerator, and FallbackWordService for enhanced game logic and error handling.
