+++ src/app/api/lexi-guess/services/WordCache.ts
